handle static resources

+++ src/server.c
@@ -13,16 +13,16 @@
 struct state GLOBAL;
 
 void set_current_page(user_t *user, char* res){
-   if (strcmp(res, "/") == 0 || strcmp(res, CLEAR_COMMAND) == 0){
+   if (res == NULL || strcmp(res, "/") == 0 || strcmp(res, CLEAR_COMMAND) == 0){
       user->current_page = INDEX_PAGE;
       return;
    }
    remove_prefix(res, "/");
-   res = strtok(res, "\r");
    if (strcmp(res, "websocket.html") == 0){
       user->current_page = GLOBAL.DEFAULT_WEBSOCKET_PAGE;
       return;
    }
+   if (strcmp(res, "post") == 0) return;
    for (int i = 0; i < LEN(available_routs); i++){
       if (strcmp(res, available_routs[i]) == 0){
          user->current_page = res;
@@ -35,11 +35,12 @@ void set_current_page(user_t *user, char* res){
 
 
 req_type handle_http_request(user_t *user, request_t *req, char* buffer, int bytes){
+   int is_static;
    char *res, *post;
    req_type type = get_type_request(buffer, bytes);
 
    req->header = "OK";
-   req->code = 200;
+   req->code = OK;
    req->is_cookie = 0;
    
    if (bytes > 0){
@@ -49,24 +50,26 @@ req_type handle_http_request(user_t *user, request_t *req, char* buffer, int byt
                ws_establish_connection(buffer, req, user);
                return WS;
             }
-            res = header_parse(buffer, bytes, " ");
-            if (strcmp(res, "/favicon.ico") == 0) {
+            res = header_parse(buffer, bytes, " ", &is_static);
+            if (strcmp(res, "favicon.ico") == 0 && is_static) {
                req->content_type = get_content_type(res);
                req->content = get_file_content("favicon.ico", &req->length, "rb");
                return OK;
-            } else if (strcmp(res, "/style.css") == 0){
+            } else if (strcmp(res, "style.css") == 0 && is_static){
                req->content_type = get_content_type(res);
                req->content = get_file_content("style.css", &req->length, "r");
                return OK;
-            }
-            else if (is_contain(res, '/')){
-               set_current_page(user, res);
-               if (strcmp(res, CLEAR_COMMAND) == 0){ //FIXME: add database
+            } else if (is_contain(res, '/') && !is_static){
+                set_current_page(user, res);
+                user->response_page = user->current_page;
+                if (strcmp(res, CLEAR_COMMAND) == 0){ //FIXME: add database
                   logger(__func__, "delete chat history");
                   system("rm -f resources/text.txt && touch resources/text.txt");
                   update_html();
                }
-            }
+
+            } else if (is_static)
+               user->response_page = res;
             break;
          case POST:
             logger(__func__, "POST request");
@@ -83,13 +86,11 @@ req_type handle_http_request(user_t *user, request_t *req, char* buffer, int byt
             return NONE;
       }
    }
-
-
-   req->content_type = get_content_type(user->current_page);
+   req->content_type = get_content_type(user->response_page);
    if (strstr(req->content_type, "image") != NULL){
-      req->content = get_file_content(user->current_page, &req->length, "rb");
+      req->content = get_file_content(user->response_page, &req->length, "rb");
    } else
-      req->content = get_file_content(user->current_page, &req->length, "r");
+      req->content = get_file_content(user->response_page, &req->length, "r");
 
    return OK;
 }
